Replace the static sponsor list with one that is dynamically created
This commit replaces the list of sponsors with a macro that I suspect will be expanded by the
markdown interpreter. The solution is adapted from the OWASP SAMM project page. The macro seems
to process the pattern from _includes/sponsors.md and fill in data from _data/sponsors.json.

The board of the German chapter decided to discontinue the old sponsor program and instead adapt
the global OWASP way of handling donations. This commit aims at autogenerating the sponsors list
from the data in _data. The hope is to get an autogenerated list of sponsors out of OWASP data
in the future.